Fixed getting score events for matches. * Endpoint requires updating in platform as matches does not call the competition.php endpoint * Changed minute from Int to String as extra time is recorded as "80+2"

+++ sport/app/rugby/feed/rugbyOptaDeserialisation.scala
@@ -87,7 +87,7 @@ object Parser {
       player.map { p =>
         ScoreEvent(
           player = p,
-          minute = (scoreEventData \ "@minute").text.toInt,
+          minute = (scoreEventData \ "@minute").text,
           `type` = ScoreType.withName((scoreEventData \ "@type").text)
         )
       }
